+++ docs/changelog.md
@@ -11,6 +11,12 @@ category: Administration
 
 ### [Unreleased]
 
+* ...
+
+### [22.01.0] - 2022-01-16
+
+**Explicitly activate and disable Scrum Poker on project-level**
+
 * docs: add developer documentation to workflow function
 * feat: verify support for Jira 8.21.0 and long-term-support releases
 * feat: verify support for latest patch releases
@@ -18,7 +24,7 @@ category: Administration
 * feat: add upgrade task to migrate from old setting to new setting
 * docs: update documentation to explain configuration options
 * feat: i18n for new configuration options on project configuration
-* docs: create archive for all 21.x releases
+* docs: create archive page for all 21.x releases
 * docs: fix table of contents for 21.x releases
 
 ### Older versions
@@ -32,4 +38,5 @@ Older versions have been moved into separate changelog documents grouped by thei
 * [Scrum Poker for Jira 2.x](/changelog-2x)
 * [Scrum Poker for Jira 1.x](/changelog-1x)
 
-[Unreleased]: https://github.com/codescape/jira-scrum-poker/compare/21.11.0...HEAD
+[Unreleased]: https://github.com/codescape/jira-scrum-poker/compare/22.01.0...HEAD
+[22.01.0]: https://github.com/codescape/jira-scrum-poker/compare/21.11.0...22.01.0